% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{available_metrics}
\alias{available_metrics}
\title{Get available performance metrics for use with other functions}
\usage{
available_metrics()
}
\value{
A vector containing the available metrics
}
\description{
Get available performance metrics for use with other functions
}
