% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{available_metrics_trial}
\alias{available_metrics_trial}
\title{Get available performance metrics}
\usage{
available_metrics_trial(loc = "metrics.R")
}
\arguments{
\item{loc}{Location of the}
}
\value{
A vector containing the available metrics
}
\description{
Get available performance metrics
}
