% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blogger_sample}
\alias{blogger_sample}
\title{Blogger.com Dataset}
\format{
## `blogger_sample`
A corpus object with 100 sample blogs.
}
\source{
<https://u.cs.biu.ac.il/~koppel/BlogCorpus.htm>

<https://huggingface.co/datasets/blog_authorship_corpus>
}
\usage{
blogger_sample
}
\description{
A subset of the dataset used by Schler and Koppel in their 2006 paper.
}
\keyword{datasets}
