% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impostor_method.R
\name{character_n_grams}
\alias{character_n_grams}
\title{Create character n-grams. n characters no spaces or words <= n characters long.}
\usage{
character_n_grams(data, n_char = 4, ...)
}
\arguments{
\item{data}{A corpus or list of corpuses}

\item{n_char}{Number of characters in character n-gram.}
}
\value{
tokenised character n-gram data
}
\description{
Create character n-grams. n characters no spaces or words <= n characters long.
}
