% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{chunk_tokens}
\alias{chunk_tokens}
\title{Chunk a Tokenised Corpus or List of Tokenised Corpuses}
\usage{
chunk_tokens(data, ...)
}
\arguments{
\item{data}{A corpus or list of corpuses}

\item{...}{Additional arguments to quanteda tokens_chunk}
}
\value{
data
}
\description{
Chunk a Tokenised Corpus or List of Tokenised Corpuses
}
