% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loaders.R
\name{extract_blog_text}
\alias{extract_blog_text}
\title{Return the blogger.com text from a filepath}
\usage{
extract_blog_text(row)
}
\arguments{
\item{row}{The row of the dataframe}
}
\value{
The date of the blogpost and the post itself
}
\description{
Return the blogger.com text from a filepath
}
