% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loaders.R
\name{lapply_parallel}
\alias{lapply_parallel}
\title{Generic lapply function to use parallel if possible}
\usage{
lapply_parallel(X, FUN, ...)
}
\arguments{
\item{X}{The list item}

\item{FUN}{The function to call}

\item{...}{Any additional function parameters}
}
\value{
A list of blogs
}
\description{
Generic lapply function to use parallel if possible
}
