% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loaders.R
\name{load_blog_corpus}
\alias{load_blog_corpus}
\title{Load Blogger.com Corpus}
\usage{
load_blog_corpus(loc, convert_to_corpus = TRUE)
}
\arguments{
\item{loc}{The location of the blog corpus}

\item{convert_to_corpus}{Binary parameter convert to quanteda corpus TRUE/FALSE}
}
\value{
The data as a corpus or dataframe
}
\description{
Load Blogger.com Corpus
}
