% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{n_grams}
\alias{n_grams}
\title{Convert Tokens into n-grams if required.}
\usage{
n_grams(data, ...)
}
\arguments{
\item{data}{A corpus or list of corpuses}

\item{...}{Additional arguments to n_gram}
}
\value{
A tokens object
}
\description{
Convert Tokens into n-grams if required.
}
