% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{remove_punctuation}
\alias{remove_punctuation}
\title{Function to remove punctuation prior to tokenising.}
\usage{
remove_punctuation(data)
}
\arguments{
\item{data}{A corpus or list of corpuses.}
}
\value{
A cleaned corpus
}
\description{
Function to remove punctuation prior to tokenising.
}
