% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{result_threshold_df}
\alias{result_threshold_df}
\title{Create a results table grouped by a threshold}
\usage{
result_threshold_df(
  data,
  start_threshold = 0,
  end_threshold = 1,
  increment = 0.01,
  metrics = c("precision", "recall"),
  unanswered_if_equal = TRUE
)
}
\arguments{
\item{data}{A dataframe to loop over various thresholds}

\item{start_threshold}{Start of the threshold options}

\item{end_threshold}{End of threshold sequence}

\item{increment}{The increments of the threshold}

\item{metrics}{The selection of metrics to use}

\item{unanswered_if_equal}{Logical value depicting whether to include unanswered option if value is the same as threshold.}
}
\value{
A result dataframe.
}
\description{
Create a results table grouped by a threshold
}
