% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{tokenise}
\alias{tokenise}
\title{Tokenise a Corpus or List of Corpus Objects.}
\usage{
tokenise(data, remove_punct_first = FALSE, ...)
}
\arguments{
\item{data}{A corpus or list of corpus object}

\item{remove_punct_first}{Remove punctuation before tokenising}

\item{...}{Additional arguments to quanteda tokens}
}
\value{
data
}
\description{
Tokenise a Corpus or List of Corpus Objects.
}
